requirements.txt
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.2
PyMySQL==1.1.1
python-dotenv==1.0.1
reportlab==4.4.5
Werkzeug==3.1.3
gunicorn==23.0.0
pillow==11.0.0
graphviz==0.20.3